Add document string $contains, $or, $and (#130)
Add support for WhereDocument type
Add validator to WhereDocument type
Plumb WhereDocument validator and as a flag to query, get, delete
Add support for string contains in clickhouse and duckdb
From stacked $and + $or support #133
Add typing for logical operators $and + $or
Add validation and validation tests for logical operators
Add support in clickhouse and duckdb. The code is duplicative across dbs but we expect this to change soon so leaving for now since footprint is minimal.