Validate index dimension, and index size (#129)
* Dim, number checks in index, plus `add` typing

* Fixing up more types

* Testing exceptions

* Good erros handing

* add pypi deploy to GH actions

* docker image code location was incorrect

* Remove newlines to improve performance

Per OpenAI's get_embedding function here: https://github.com/openai/openai-python/blob/main/openai/embeddings_utils.py

* Add `latest` tag to docker releases

For tagged releases, add the `latest` tag to the Docker image as well.

* Add document string $contains, $or, $and (#130)

Add support for WhereDocument type
Add validator to WhereDocument type
Plumb WhereDocument validator and as a flag to query, get, delete
Add support for string contains in clickhouse and duckdb
From stacked $and + $or support #133
Add typing for logical operators $and + $or
Add validation and validation tests for logical operators
Add support in clickhouse and duckdb. The code is duplicative across dbs but we expect this to change soon so leaving for now since footprint is minimal.

* Remove uuid since it is a part of the python std lib (#136)

* Remove uuid since it is a part of the python std lib
* Remove from pyproject.toml

* Add delete tests back and where_document null handling (#137)

* Fix up tests

* Add document string $contains, $or, $and (#130)

Add support for WhereDocument type
Add validator to WhereDocument type
Plumb WhereDocument validator and as a flag to query, get, delete
Add support for string contains in clickhouse and duckdb
From stacked $and + $or support #133
Add typing for logical operators $and + $or
Add validation and validation tests for logical operators
Add support in clickhouse and duckdb. The code is duplicative across dbs but we expect this to change soon so leaving for now since footprint is minimal.

* Add delete tests back and where_document null handling (#137)

* Fix up nulls, handle server responses

---------

Co-authored-by: Luke VanderHart <luke@vanderhart.net>
Co-authored-by: Daniel Gross <daniel@dcgross.com>
Co-authored-by: Hammad Bashir <HammadB@users.noreply.github.com>