Fix embedding function (#156)
Fix a bug where persistence and embedding_functions are fundamentally incompatible. We do not allow a user to get_collection and supply an embedding function. Previously, if they were to create_collection with an embedding_function and then create a new client, get_collection would not be aware of the embedding_function, leading to it use the default. This change allows get'ing a collection and allowing for an embedding_function to be passed.