Add support for multiple spaces (#457)
Add support for multiple distance metrics in tests.

We coin-flip and sometimes add a space when using hnsw_params
Added the distance functions to the invariant and use them when needed.
In the process of writing this test I discovered a bug with our implementation of update that was revealed by the inner product space. Since the inner product is not a true metric, a point may not be a neighbor to itself. Our update code was strictly appending to the index due to the a bug with how we manage string UUID vs UUID objects. In l2 and cosine spaces, this usually was fine in the eyes of tests since the results returned were correct with the updated data. But IP exacerbated the issue by making the results not always be the same point.