Change cross version tests to use standard import from path  (#557)
Change cross version tests to use standard import from path as opposed
to importlib

## Description of changes
In #541 we began to import settings as chromadb.config.Settings. This
broke cross version tests after the release of v0.3.23. Due to the
import changing for settings in LocalAPI's constructor, it highlighted
an issue with how we import modules in cross version tests. Namely,
nested modules are not imported unless explicit. When crossing the
process boundary, we don't import this module and so it breaks when
referred to. We now instead use standard python imports instead which
works because of the path being overridden. This is also generally
simpler and better.

## Test plan
Existing tests.

I verified that a manually introduced bug (same as #386) breaks the
tests.
I printed sys.modules and verified that all paths are correct.

## Documentation Changes
None required.