switch to named params for JS client (#513)
In this PR
- docstrings! 
- types!
- named params! 

This is a breaking change to the JS Client - it switches from positional
to named arguments originally brought up by @transitive-bullshit in
https://github.com/chroma-core/chroma/issues/254

Example: 

```
// before
const collection = await chroma.createCollection("test");

//after
const collection = await chroma.createCollection({ name: "test" });
```

This is especially nice when you are only using a few params. 

This PR does not bump the `npm` version number - we will cut a new
release on Monday.

We will need to land an update to our docs at that time as well. TBD