Add task instruction pairing to InstructorEmbeddingFunction (#556)
## Description of changes

- Add an optional `instruction` constructor parameter to
InstructorEmbeddingFunction to allow `instruction` and Document pairs to
be encoded.

## Test plan


## Documentation Changes
Added examples to the Alternative Embedding notebook.

Not sure if this is a good implementation, since you'll need a separate
Collection for each instruction you want to use (or reassign
`self._instruction`), but at least the change is pretty minimal. For my
use case, two instructions are enough (one for storing, one for
retrieving). For a scenario where you need lots of different
instructions, perhaps "Represent the <Science|Financial|Political|etc.>
article: ", another solution is needed.

Feature Request #546

---------

Co-authored-by: Jeffrey Huber <jeff@trychroma.com>