Web AI embedder for JS library (#633)
## Description of changes

Added a new JavaScript embedding function to generate embeddings for
text and images using the [Web AI](https://github.com/visheratin/web-ai)
library. The users will now be able to generate the embeddings in both
NodeJS and (IMO more importantly) in the browser. This, along with the
existing JS library design, enables the decoupling of embeddings
generation from storage (see example demo below).

## Test plan

I created a demo to check how the function works in web apps -
https://web-ai-chroma.vercel.app/

The embeddings are generated on the client side and sent to the server
to add to the collection or query for similar items.

## Documentation Changes

Likely this change will require changes in the documentation.