Fix multi-add property test for SQLite (#682)
## Description of changes

The SQLite implementation does not emit an exception to the API when
there are duplicate IDs on `add`. This updates the tests to reflect
that.

In future, we should instead consider making upsert and delete the only
API operations.

## Test plan
This is the test 

## Documentation Changes
None, however we should update the documentation when this branch lands.