Fix sqllite filtering (#692)
## Description of changes
Sqlite treats _ as a wildcard operator and % as one or many operator.
When using LIKE we have to escape these characters and then use the
ESCAPE keyword. This PR adds a custom pypika function for doing escaped
like queries.

Additionally, sqllite stops processing strings after the null (\x00)
character. We should not generate this in our tests, and users should
likely sanitize this character.

## Test plan
These are fixes to make tests pass.

## Documentation Changes
None required.