Enable custom JS Client headers during request to Chroma Instance (#728)
connect #725

## Description of changes
Add a new `fetchOptions` parameter on the JS ChromaClient creation that
allows you to pass in custom `fetch` options.

Example:
```
const chroma = new ChromaClient({
  path: "http://localhost:8000",
  fetchOptions: {
    headers: {
      'X-Api-Token': "sk-live-Hunt3r2", // Works like regular node-fetch headers!
    }
  }
});
```

## Test plan
The existing test suite was run and validated as passing. 

## Documentation Changes
Typescript would allow the developer to know that these options are
available. I do not think it currently requires updates to the Chroma
Documentation - especially since this change is only impacting the JS
client.

---------

Co-authored-by: Jeffrey Huber <jeff@trychroma.com>