Normalized embeddings (#737)
## Description of changes
Added support to the use of normalize embeddings in the
`SentenceTransformerEmbeddingFunction` class by adding a new attribute
to the class and use it when calling the `encode` function.

## Test plan
I was planning to add a new test case where an object was created with
`normalize_embeddings=True` but I haven't found any test which is
currently testing `SentenceTransformerEmbeddingFunction` (I guess that
should be in the folder `test/utils`).

## Documentation Changes
The document talking about [embeddings] should be changed. In the
section *Sentence Transformers* it should be mentioned the possiblitiy
of using an optional parameter `normalize_embeddings`. A similar text to
the one in the SentenceTransformer documentation could be used (or
adapted):

> If set to True, returned vectors will have length 1. In that case, the
faster dot-product (util.dot_score) instead of cosine similarity can be
used.