Fix empty metadata, partial updates.  (#699)
## Description of changes

*Summarize the changes made by this PR.*
1. The new sqlite refactor cannot represent None and the empty dict in
metadata. We make the choice here to change all of our code to validate
against the empty dict for metadata. The metadata can now only be None
or a non-empty dictionary. We add validation for this and modify the old
and new clients to match the behavior. These changes affect both
collection and embedding metadata.
2. The new API allows partial updates for existing ID's. If we add 1,2,3
and then add 1,2,3,4. 4 will still get added while 1,2,3 will silently
fail. We modify the old API to match this behavior.
3. The new API will merge metadata on upsert, the tests special case
this for the new api as porting this to the old api is not worth the
investment at this time.
4. The Sqlite code was preventing falsy (such as 0) metadata values from
being read. This addresses that.

## Test plan
These are all tested via old or modified tests.

## Documentation Changes
We should update the docs to discuss:
- Partial success
- Merging of metadata in update/upsert