[Experimental] Distance distribution relevance (#762)
## Description of changes

This PR introduces the `experimental/` subdirectory, which is intended
to contain prototypes and experiments of various ideas and functionality
related to embedding based retrieval, and retrieval augmented
generation.

In the subdirectory is a notebook which illustrates the idea of using a
distribution of distances within a dataset to compute a metric for
relevance, which is invariant to the embedding function and data itself.

## Test plan
None. 

## Documentation Changes
None, Notebook is self-documenting. We may want to add an Experimental
page on the documentation site in future.