cross compatible fetch with generation pipeline integration (#752)
Related issue #684 
## Description of changes
 - Improvements & Bug fixes
   - npm client is now cross compatible for Node versions <18. 
 - New functionality
- I tested 4 versions: 18.16.0 PASS, 16.18.1 PASS, 14.21.3 PASS, 10.24.1
FAIL

## Test plan
The same tests in
[root/client/js/test](https://github.com/chroma-core/chroma/tree/main/clients/js/test)
must successfully pass under different Node versions for this to work
(they did pass).

## Documentation Changes
Perhaps add here https://docs.trychroma.com/api-reference the supported
version range.

## Notes
The core of this improvement is incredibly simple, it boils down to
adding
```ts
import 'isomorphic-fetch';
```
to
[runtime.ts](https://github.com/chroma-core/chroma/blob/main/clients/js/src/generated/runtime.ts).

Due to the fact that `openapi-generator-plus` is being used I included
some commands to the
[genapi.sh](https://github.com/chroma-core/chroma/blob/main/clients/js/genapi.sh)
to ensure `isomorphic-fetch` will be imported correctly whenever the api
schema is regenerated.

Unfortunately, the openapi-generator has not been run for a while
leading me to believe there is discontinuity between the typescript
package and server openapi typings. As a result running
[genapi.sh](https://github.com/chroma-core/chroma/blob/main/clients/js/genapi.sh)
(running openapi-generator) will lead to clashes in typings. I have
created the following issue #751 reguarding this