Remove non-sqlite impls and code. Add sqlite <> fastapi tests. Add sqlite connection pool. (#781)
## Description of changes
1. Removes the old non-sqlite impls and associated code.
2. The fastapi -> sqlite code path was untested. This added that test
config, and caught an issue with sqlite not supporting passing the
connection between threads. Since fastapi runs our code in a thread
pool, this was broken.
3. Add a connection pool for sqlite that creates a unique conn per
thread, sqlite connections cannot be passed across threads.
4. Add a legacy setting detector (Apologies for overloading this PR)
5. By default, sqlite does not support threads sharing memory, however
this is possible by using "file::memory:?cache=shared" instead of
":memory:" setting.


## Test plan
Existing tests were instrumented for this code.
 
## Documentation Changes
We need to update all website code to be accurate for sqlite.