Make migrations system use importlib resources (#810)
## Description of changes
The migrations system was using project-relative paths to import the
migration sql's into code. This does not work well with packaging chroma
for distribution, so we opt to move migrations into the package and use
importlib_resources which is the recommended way to bundle files with a
package
https://setuptools.pypa.io/en/latest/userguide/datafiles.html#accessing-data-files-at-runtime.

## Test plan
Existing tests cover these changes.

## Documentation Changes
None required.