Add docs, and minor logging fix (#820)
## Description of changes
Updates various docs and examples to be relevant for the sqlite
refactor, adds a small QoL fix to prevent a confusing log.

1. Update docker-entrypoint.sh to point to chroma-hnswlib
2. Update docker-compose in bin/templates to not use clickhouse
3. Update local_persistent_hnsw to not overquery underlying index past
its size as it will result in a confusing double log. This doesn't
modify the logic as the underlying index was re-performing this logic
before. But it led to a confusing log where we repeatedly told you we
were changing your query
4. Update the thin client readme to use HttpClient
5. Make the dockerfile and compose files use IS_PERSISTENT=TRUE since
its false by default and the default client/server path should be
persistent.
6. Update all notebooks that used the old client by modifying +
rerunning


## Test plan
Existing tests, or they are notebooks which I ran

## Documentation Changes
None