Revert trigram tokenizer (#841)
## Description of changes

Because of inconsistent bundling, not all python versions on all
platforms ship with a SQLite which has the trigrams tokenizer, and the
same python version may have different SQLites on different platforms.

This creates problems for users of Chroma `0.4.0` which uses the trigram
tokenizer. This change reverts to using the default tokenizer of
whatever SQLite is available on the system. The tradeoff is reduced
full-text search performance, but this simplifies installation and
maintenance.


## Test plan
The test suite should pass. 

## Documentation Changes
None, not user-facing.