New starter notebook (#881)
## Description of changes

This PR creates a new starter notebook intended to familiarize people
with the very basic, core functionality of embedding retrieval with
Chroma. It's self-contained, and hopefully straightforward and easy to
understand.

There is also a minor fix to the experimental notebook. 

## Test plan

Ran the notebook, also via Colab. 

## Documentation Changes
None. 

## TODO
- [x] https://github.com/chroma-core/chroma/issues/880 Canonical 'chat
with your documents'