fix: Unziping the onnyx model tarball no longer depends on the archive noe being present. This is useful in situations where the tarball was downloaded but never extracted. One side effect is that we'll always unzip the tarball but then we don't make the assumption that everything is working.
Side note on security:  We could check checksum of the file, can also be hosted in s3. Just to prevent possible attack vector where model is replaced locally on the machine. An added benefit would be that if there are new versions of the model where checksum will not match we can re-download and extract or at least warn the user about new version.