Grpc Segments + Distributed Segment Manager (#952)
## Description of changes
*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - N/A
 - New functionality
	 - Adds gRPC based segments backed by local segment impls
	 - Add protobuf service defs for the distributed index type
	 - Extend convert.py with the needed proto conversions
	 - Adds crude distributed segment manager
	 - For now, only get() is supported, query is not
	 - For now only vector segments are supported

## Test plan
This PR adds a test harness to run the tests against the distributed
impl. However the tests assume synchronous behavior in many places, and
need to be overhauled to support the new decoupled read/write path.
Rather than bloat this PR, we add the tests as a new GH workflow (Chroma
Cluster Tests), expect them to fail, and will add a subsequent PR to
both add the NotImplemented() functionality as well as fix the tests
gradually over time. (The tests are green but if you drill in they are
failing, this is the closest github actions gets to xfail)

As we develop this, the goal is to make the tests progressively pass,
rather than bloat PRs until we are at feature parity.

How to test locally:
Run `bin/cluster-test.sh`

## Documentation Changes
N/A for now