[STACKED #1255] [ENH] Add multitenancy (#1244)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - ...
 - New functionality
- Adds multitenancy and databases as first class concepts by migrating
the db and plumbing it through the API into the sysdb
- We now treat the "System" as a singleton-per-path and create a wrapper
API object that proxies to it with context on the tenant/database. In
this way the server is context-unaware about the connection.

## Test plan
*How are these changes tested?*
Unit Tests were added for new client tenant/database behavior
Property tests were added for the new tenant/database behavior by
subclassing the collection state machine and switching the
tenant/database as a state machine transition.

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js

## Documentation Changes
I will add a section to the docs about multitenancy and how to use it.
We can remove warnings about the client being a singleton.