[ENH] URI Data Loader (#1294)
## Description of changes

This PR adds URIs and DataLoaders into Chroma. 
- `DataLoader` works like `EmbeddingFunction`, except it takes a `URIs`
and outputs the specified datatype.
- A `DataLoader` using `pillow` for image file loading. 
- Adds `uris` as a field on `add`, `query`, as well as an `include`
field
- Adds `data` as an `include` field 

URIs specify a place where data can be loaded from, and can be used to
load data for embedding, or as the result of retrieval.

This makes multimodal retrieval with data stored externally as files
seamless and extensible.

This PR is stacked on https://github.com/chroma-core/chroma/pull/1293

## Test

Integration tests pass. 

A new unit test for data loaders:
https://github.com/chroma-core/chroma/blob/c71c3efa15d2a9252db26470b9ff1cdb10a2b681/chromadb/test/data_loader/test_data_loader.py

Try the notebook:
https://github.com/chroma-core/chroma/blob/c71c3efa15d2a9252db26470b9ff1cdb10a2b681/examples/multimodal/multimodal_retrieval.ipynb

## Documentation

Documentation for this and
https://github.com/chroma-core/chroma/pull/1293
https://github.com/chroma-core/docs/pull/157

## TODOs
- [x] Concurrent Loading
- [x] Tests
- [x] Wiring through FastAPI
- [x] Documentation