[ENH] Update SubmitEmbeddingRecord to take collection_id. Update EmbeddingRecord to take collection_id (#1359)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes

In the distributed architecture, collections are multiplexed onto
topics, so messages from the ingest need to be tagged with their
collection_id so consumers know which collection it belongs to. In the
single node collections are 1:1 with topics so this mapping is not
strictly needed since the topic implicitly tells consumers which
collection they are reading. This PR adds the collection id to the type
that gets written on the log. It also adds an optional collection_id to
the type consumed from the log. This must be optional for backwards
compatibility on single node chroma. A backfill would have no data to
populate that field, and rather than populate it with a bad placeholder,
I think its better to be specific that it may not be present and let
downstream code decide how to handle that.

 - New functionality
	 - None.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js

## Documentation Changes
None required.