[ENH] Cloud client (#1462)
## Description of changes

This PR adds the Chroma `CloudClient` as a client type. It takes an API
key (or reads it from the environment, or prompts the user for it if
it's not found) , and tries to connect to the Chroma cloud using the API
key.

Under the hood, this is a specialized HTTPClient with ssl always turned
on, and which uses token-based authentication with the API key as the
token.

This PR also moves the `AuthorizationError` from auth to the general set
of errors, so it can be handled elsewhere by the Chroma system.

## Test plan

Added a new test, `test_cloud_client`, which creates a mock 'cloud'
server and then tests connection with valid and invalid API keys.

## Documentation Changes

Cloud is not yet available, so `CloudClient` is for now not to be used
by external users of Chroma. Therefore, we won't land documentation
changes with this PR.

## TODOs
- [x] Tests
- [ ] JavaScript @jeffchuber to take a look
- [ ] ~Docs~