[ENH] Add Notifications for CreateCollection and DeleteCollection (#1446)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - ...
 - New functionality
- This PR uses outbox pattern to record CreateCollection and
DeleteCollection events in a state store (memory or database) and then
push the notifications with notifiers. The goal is to notify frontend
servers and indexing workers to perform corresponding actions such as
disallow queries on deleted collection.
- Note that there are quite a few follow ups and optimizations. 1. Make
the notification semi sync for better consistency guarantee 2. For other
updates, there can be cases that the frontend servers and indexing
servers have more up to date metadata and do not need to apply updates
for previous versions. We can introduce version to the collection to
handle those cases.

## Test plan
*How are these changes tested?*

- [ ] Unit tests in golang
- [ ] Manual integration testing

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*