[ENH] Add rust rendezvous hashing and errors (#1508)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - /
 - New functionality
- This adds the rendezvous hashing to the rust code so that it may be
used by assigners.
- It also introduces our philosophy on error handling in the rust
codebase. We will use the 17 standard error codes in the grpc codebase,
this is used so lower levels can send their codes higher, and in higher
levels of the code we can Box<> a dyn Chroma Error to allow for generic
handling over the codes. Lower levels will have to be specific, at the
expense of verbosity in the errors they propagate and wrap. We will use
the thiserror library, widely adopted in rust, in order to easily derive
our errors. If a type needs to wrap a source error, thiserror provides
#[from] in order to allow callers to have more nuanced information.
After much research, my philosophy was informed by this post -
https://mmapped.blog/posts/12-rust-error-handling.html.

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `cargo test`

## Documentation Changes
None required. Please evaluate the documentation quality of the code
itself.