CLN: Updated versions of `flake8` and `pre-commit-hooks` and modified files accordingly (#1575)
## Description of changes
 **Improvements & Bug fixes**
- Currently, most of the dependencies in `.pre-commit.config.yaml` file
are not upto date. So, I updated them
(`flake8` and `pre-commit-hooks`) to their latest versions. Also, the
most of the pre-commit hooks are failing. So, i fixed the following:
	 1. `trailing-whitespace` hook by running `pre-commit run --all-files`.
	 2. `fix end of files` hook by running `pre-commit run --all-files`.
3. Fixed some flake8 errors like `E721 : Do not compare types use
isinstance()`.
4. Added `F401 : module imported but unused` rule in extend-ignore
because it's failing in a quite a few places and as it is not so
important rule.
	
- New functionality
	 - ...

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js

## Documentation Changes
No need of any documentation changes.