[ENH] Upgrade tests and release to Python 3.12 (#1715)
## Description of changes

Chroma did not support Python 3.12 because of our dependency on the ONNX
runtime for our default embedding function. As of version 1.17.0, ONNX
supports python 3.12:
https://github.com/microsoft/onnxruntime/issues/17842#issuecomment-1936484800

This already automatically fixes the issue for Chroma users when they
install the new version of ONNX / reinstall Chroma. This PR is just to
update our test and release actions to also use python 3.12.

## Test plan

These are changes to test workers. 

## Documentation Changes
N/A