[BUG] Log service incorrectly doesn't hydrate collection id (#1922)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Log service incorrectly did not return the collection ID of the
record. I suspect that we meant to strip the collection id and rehydrate
it into the proto for space reasons. I honored this intent and rehydrate
the colleciton id.
	 - Add some readme for running tests locally with PG
- Fixed a bug in the log_service test where the input is mutated, which
makes the source of truth have no collection id. This was passing when
we incorrectly returned no collection id but was correctly failing now.
I patched the test by cloning the records for a SOT.
	 - For the test I fixed, the expected vs actual order was incorrect.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None