[CLN] Restructure EmbeddingRecord -> LogRecord. Use 'log_offset' as term instead of 'id' (#1934)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- This PR follows #1933 and restructures EmbeddingRecord to be a nested
type around OperationRecord. We rename EmbeddingRecord to be LogRecord.
A LogRecord is a model of an OperationRecord stored on a log, and just
keeps a log_offset The log is assumed to be per-collection.
- Rename the Logservices use of "ID" to instead be "log_offset" to
standardize on the "log_offset" terminology.
- Rename RecordLog -> LogRecord for conceptual clarity and to align with
other types.
- Fixes a proto style guide violation where we used camelcase for
fieldnames -
https://protobuf.dev/programming-guides/style/#message_and_field_names.
- In GetColllectionCompactionInfo rename first_log_id to
first_log_offset - which is much clearer. Also rename first_log_id_ts to
first_log_offset. What is a id_ts? Confusing name.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None